cleaning code and adding documentation

+++ server/server.js
@@ -17,30 +17,29 @@ var app = express();
 
 app.use(bodyParser.json());
 
-app.post('/todos', authenticate, (req, res) => {
+app.post('/todos', authenticate, async (req, res) => {
 	var todo = new Todo({
 		text: req.body.text,
 		_creator: req.user._id
 	});
-
-	todo.save().then((doc) => {
-		res.send(doc);
-	}, (e) => {
+	try {
+		var doc = await todo.save()
+		res.send(doc)
+	} catch (e) {
 		res.status(400).send(e);
-	});
+	}
 });
 
-app.get('/todos', authenticate, (req, res) => {
-	Todo.find({
-		_creator: req.user._id
-	}).then((todos) => {
+app.get('/todos', authenticate, async (req, res) => {
+	try {
+		var todos = await Todo.find({_creator: req.user._id})
 		res.send({todos})
-	}, (e) => {
-		res.status(404).send(e);
-	});
+	} catch (e) {
+		res.status(404).send(e);		
+	}
 });
 
-app.get('/todos/:id', authenticate, (req, res) => {
+app.get('/todos/:id', authenticate, async (req, res) => {
 	var id = req.params.id;
 
 	if (!ObjectID.isValid(id)) {
@@ -48,20 +47,25 @@ app.get('/todos/:id', authenticate, (req, res) => {
 		return console.log(`${id} is not a valid id`)
 	}  
 	
-	Todo.findOne({
+	try {
+	var todo = await Todo.findOne({
 		_id: id,
 		_creator: req.user._id
-	}).then((todo) => {
-		if (!todo) {
+	})
+	
+	if (!todo) {
 			res.status(404).send()
 			return console.log('ID not found')
-		}
-		res.send({todo})
-		res.status(200).send()
-	}).catch((e) => console.log(e));
+	}
+
+	res.send({todo})
+	res.status(200).send()
+	} catch (e) {
+		res.status(400).send()
+	}
 });
 
-app.delete('/todos/:id', authenticate, (req, res) => {
+app.delete('/todos/:id', authenticate, async (req, res) => {
 	var id = req.params.id;
 
 	if (!ObjectID.isValid(id)) {
@@ -69,22 +73,22 @@ app.delete('/todos/:id', authenticate, (req, res) => {
 		return console.log(`${id} is not a valid id`)
 	}  
 
-	Todo.findOneAndRemove({
-		_id: id,
-		_creator: req.user._id
-	}).then((todo) => {
+	try {
+		var todo = await Todo.findOneAndRemove({_id: id, _creator: req.user._id})
+		
 		if(!todo) {
 			res.status(404).send();
 			return console.log('no todo found with that ID');
 		}
 
 		res.send({todo});
-	}).catch((e) => {
+
+	} catch (e) {
 		res.status(400).send();
-	});
+	}
 });
 
-app.patch('/todos/:id', authenticate, (req, res) => {
+app.patch('/todos/:id', authenticate, async (req, res) => {
 	var id = req.params.id;
 	// only allows the text and completed fields to be modified by the endpoint
 	var body = _.pick(req.body, ['text', 'completed']);
@@ -99,68 +103,68 @@ app.patch('/todos/:id', authenticate, (req, res) => {
 		body.completed = false;
 		body.completedAt = null;
 	}
-
-	// new:true returns the new doc rather than the old one
-	Todo.findOneAndUpdate({
+	try {
+	var todo = await Todo.findOneAndUpdate({
 		_id: id,
 		_creator: req.user._id
-	}, {$set: body}, {new:true}).then((todo) => {
-		if (!todo) {
+	}, {$set: body}, {new:true})
+	
+	if (!todo) {
 			return res.status(404).send();
-		}
+	}
 
-		res.send({todo});
-	}).catch((e) => {
+	res.send({todo});
+
+	} catch (e) {
 		res.status(400).send();
-	})
+	}
 });
 
-app.post('/users', (req, res) => {
-	var body = _.pick(req.body, ['email', 'password']);
-	var user = new User(body);
-
-	user.save().then(() => {
-		// instead of responding, will call new instance method instead
-		// res.send(user);
-		return user.generateAuthToken();
-	}).then((token) => {
+app.post('/users', async (req, res) => {
+	try {
+		var body = _.pick(req.body, ['email', 'password']);
+		var user = new User(body);
+		await user.save();
+		var token = await user.generateAuthToken();
 		res.header('x-auth', token).send(user);
-	}).catch((e) => {
+	} catch (e) {
 		res.status(400).send(e);
-	});
+	}
 });
 
-app.get('/users', (req, res) => {
-	User.find().then((users) => {
+app.get('/users', async (req, res) => {
+	
+	try {
+		var users = await User.find()
 		res.send({users})
-	}, (e) => {
+	} catch (e) {
 		res.status(404).send(e);
-	});
+	}
 });
 
 app.get('/users/me', authenticate, (req, res) => {
 	res.send(req.user);
 });
 
-app.post('/users/login', (req, res) => {
-	var body = _.pick(req.body, ['email', 'password']);
-
-	User.findByCredentials(body.email, body.password).then((user) => {
-		return user.generateAuthToken().then((token) => {
-			res.header('x-auth', token).send(user);
-		});
-	}).catch((e) => {
+app.post('/users/login', async (req, res) => {
+	try {
+		var body = _.pick(req.body, ['email', 'password']);
+		var user = await User.findByCredentials(body.email, body.password);
+		var token = await user.generateAuthToken();
+		res.header('x-auth', token).send(user);
+	} catch (e) {
 		res.status(400).send();
-	});
+	}
 });
 
-app.delete('/users/me/token', authenticate, (req, res) => {
-	req.user.removeToken(req.token).then(() => {
-		res.status(200).send();
-	}, () => {
+app.delete('/users/me/token', authenticate, async (req, res) => {
+	try {
+		await req.user.removeToken(req.token);
+		res.status(200).send()
+	} catch (e) {
 		res.status(400).send()
-	});
-})
+	}
+});
 
 
 app.listen(port, () => {
